refactor(reservation): 네이밍 수정 및 JOIN 로직 추가
