--- src/main/java/roomescape/reservation/QueryingDAO.java
@@ -1,30 +1,0 @@
-package roomescape.reservation;
-
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.core.RowMapper;
-import org.springframework.stereotype.Repository;
-
-import java.util.List;
-
-@Repository
-public class QueryingDAO {
-    private JdbcTemplate jdbcTemplate;
-
-    public QueryingDAO(JdbcTemplate jdbcTemplate) {
-        this.jdbcTemplate = jdbcTemplate;
-    }
-
-    private final RowMapper<Reservation> reservationRowMapper = (resultSet, rowNum) -> {
-        return new Reservation(
-                resultSet.getLong("id"),
-                resultSet.getString("name"),
-                resultSet.getDate("date").toLocalDate(),
-                resultSet.getTime("time").toLocalTime()
-        );
-    };
-
-    public List<Reservation> findAll() {
-        String sql = "select id, name, date, time from reservation";
-        return jdbcTemplate.query(sql, reservationRowMapper);
-    }
-}
