+++ src/main/java/roomescape/reservation/ReservationQueryingDao.java
@@ -1,0 +1,61 @@
+package roomescape.reservation;
+
+import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.jdbc.core.RowMapper;
+import org.springframework.stereotype.Repository;
+import roomescape.reservationTime.ReservationTime;
+
+import java.time.LocalTime;
+import java.util.List;
+
+@Repository
+public class ReservationQueryingDao {
+    private JdbcTemplate jdbcTemplate;
+
+    public ReservationQueryingDao(JdbcTemplate jdbcTemplate) {
+        this.jdbcTemplate = jdbcTemplate;
+    }
+
+    private final RowMapper<Reservation> reservationRowMapper = (resultSet, rowNum) -> {
+        ReservationTime time = new ReservationTime(
+                resultSet.getLong("time_id"),
+                LocalTime.parse(resultSet.getString("start_at"))
+        );
+
+        return new Reservation(
+                resultSet.getLong("id"),
+                resultSet.getString("name"),
+                resultSet.getDate("date").toLocalDate(),
+                time
+        );
+    };
+
+    public List<Reservation> findAll() {
+        String sql = """
+        SELECT 
+            r.id as id, 
+            r.name as name, 
+            r.date as date, 
+            t.id as time_id, 
+            t.start_at as start_at 
+        FROM reservation r 
+        INNER JOIN reservation_time t ON r.time_id = t.id
+        """;
+        return jdbcTemplate.query(sql, reservationRowMapper);
+    }
+
+    public Reservation findById(Long id) {
+        String sql = """
+        SELECT
+            r.id as id,
+            r.name as name,
+            r.date as date,
+            t.id as time_id,
+            t.start_at as start_at
+        FROM reservation r
+        INNER JOIN reservation_time t ON r.time_id = t.id
+        WHERE r.id = ?
+        """;
+        return jdbcTemplate.queryForObject(sql, reservationRowMapper, id);
+    }
+}
